% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DiagnosisCleaner}
\alias{DiagnosisCleaner}
\title{DiagnosisCleaner}
\usage{
DiagnosisCleaner(DT)
}
\arguments{
\item{DT}{a data table containing a column named "patientdiagnosis" and a column named "recno".}
}
\value{
a data table containing one column labeled "diagnoses", a second colum labeled "annotation" containing the diagnosis annotation, and a third column labeled "recno" containing the initial record number the diagnosis originated from.
}
\description{
separates out the diagnosis code from the diagnosis annotation.
}
\author{
Brendan Gongol
}
