% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{EmptyColumnRemoveR}
\alias{EmptyColumnRemoveR}
\title{EmptyColumnRemoveR}
\usage{
EmptyColumnRemoveR(DT)
}
\arguments{
\item{DT}{the data table to remove the empty columns from.}
}
\value{
a data table containing conly columns that have at least one entry.
}
\description{
Removes the empty columns from a data table.
}
\author{
Brendan Gongol
}
