% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{frequenciesdyn}
\alias{frequenciesdyn}
\title{frequenciesdyn}
\usage{
frequenciesdyn(DTstr, xstr)
}
\arguments{
\item{DTstr}{the objest of a data table entered in quotations (eg: "ulcer")}

\item{xstr}{a column in a data table entered in quotations (eg: "patmrnid")}
}
\value{
frequency counts of xstr in DTstr
}
\description{
Computes the frequencies of a particular attribute
}
\author{
Brendan Gongol
}
