% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{labAnnotatR}
\alias{labAnnotatR}
\title{Labcleanr}
\usage{
labAnnotatR(DT, type)
}
\arguments{
\item{DT}{a data table containing a column named "low" and a column named "high". If type = "prior", a column labeled "labs_prior" and a column labeled "values_prior", if type = "post", a column labeled "labs_after" and a column labeled "values_after".}

\item{type:}{either prior of post}
}
\value{
a vector with annotated lab values
}
\description{
transforms lab values from a continuous to categorical variable.
}
\author{
Brendan Gongol
}
