% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Labcleanr}
\alias{Labcleanr}
\title{Labcleanr}
\usage{
Labcleanr(DT, type)
}
\arguments{
\item{DT}{a data table containing a column named "recno". If type = "prior", a column labeled "labsprior", if type = "post", a column labeled "labsafter"}

\item{type:}{either prior of post}
}
\value{
a data table containing one column labeled "labs_prior", a column labeled "values_prior", and a third colum labeled "recno" containing the initial record number the diagnosis originated from.
}
\description{
cleans the lab value annotation in a data table column.
}
\author{
Brendan Gongol
}
