% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{loaddata}
\alias{loaddata}
\title{loaddata}
\usage{
loaddata(fname)
}
\arguments{
\item{fname}{a csv file to load.}
}
\value{
A loaded data.table
}
\description{
loads the data for use in further analyses.
}
\author{
Brendan Gongol
}
