% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{MedsCleanR}
\alias{MedsCleanR}
\title{MedsCleanR}
\usage{
MedsCleanR(DT, type)
}
\arguments{
\item{DT}{a data table containing a column named "recno". If type = "prior", a column labeled "medsprior", if type = "after", a column labeled "medsafter"}

\item{type:}{either prior of post}
}
\value{
a data table containing one column labeled "labs_prior", a column labeled "values_prior", and a third colum labeled "recno" containing the initial record number the diagnosis originated from.
}
\description{
cleans the medications column of a data table.
}
\author{
Brendan Gongol
}
