% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ProcedureCleaner}
\alias{ProcedureCleaner}
\title{ProcedureCleaner}
\usage{
ProcedureCleaner(DT)
}
\arguments{
\item{DT}{a data table containing a column named "patientdiagnosis" and a column named "recno" containing the record for each record.}
}
\value{
a data table containing one column labeled "orprocedures" and a second colum labeled "recno" containing the initial record number the diagnosis originated from.
}
\description{
cleans the OR procedures into a usable format.
}
\author{
Brendan Gongol
}
