% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{PUscoreR}
\alias{PUscoreR}
\title{PUscoreR}
\usage{
PUscoreR(DT, diag, proc, lab, med)
}
\arguments{
\item{DT}{a data table containing labels specified according to the diagnosis, procedure, lab, or meds cleaning functions.}

\item{diag}{either TRUE or FALSE}

\item{proc}{either TRUE or FALSE}

\item{lab}{either TRUE or FALSE}

\item{med}{either TRUE or FALSE}
}
\value{
a vector with scored values
}
\description{
scores pressure ulcer severity.
}
\author{
Brendan Gongol
}
