% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{StageCleanR}
\alias{StageCleanR}
\title{EmptyColumnRemoveR}
\usage{
StageCleanR(DT)
}
\arguments{
\item{DT}{a data table containing a column labeled "pressureulcerstaging".}
}
\value{
a data table containing the pressure ulcer stages according to record number
}
\description{
Removes the empty columns from a data table.
}
\author{
Brendan Gongol
}
